+++ confirm.php
@@ -24,6 +24,11 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
+use local_booking\local\message\notification;
+use local_booking\local\participant\entities\student;
+use local_booking\local\session\entities\booking;
+use local_booking\local\subscriber\entities\subscriber;
+
 require_once(__DIR__ . '/../../config.php');
 require_once(__DIR__ . '/lib.php');
 
@@ -37,7 +42,40 @@ $context = context_course::instance($courseid);
 require_login($courseid, false);
 require_capability('local/booking:availabilityview', $context);
 
-list($result, $time, $week) = confirm_booking($courseid, $instructorid, $studentid, $exerciseid);
+$COURSE->subscriber = new subscriber($courseid);
+$result = -1;
+$time = time();
+$week = (int) date('W', time());
+
+// Get the student slot
+$booking = new booking(0, $courseid, $studentid, $exerciseid);
+$booking->load();
+
+if (!empty($booking->get_id())) {
+    // update the booking by the instructor.
+    $sessiondatetime = (new DateTime('@' . ($booking->get_slot())->get_starttime()))->format('D M j\, H:i');
+    $strdata = [
+        'exercise'  => $COURSE->subscriber->get_exercise_name($exerciseid),
+        'instructor'=> student::get_fullname($instructorid),
+        'status'    => ucwords(get_string('statusbooked', 'local_booking')),
+        'sessiondate'=> $sessiondatetime
+    ];
+    if ($booking->confirm(get_string('bookingconfirmmsg', 'local_booking', $strdata))) {
+        // notify the instructor of the student's confirmation
+        $message = new notification();
+        $result = $message->send_instructor_notification($courseid, $studentid, $exerciseid, $sessiondatetime, $instructorid);
+    }
+    $time = ($booking->get_slot())->get_starttime();
+    $week = ($booking->get_slot())->get_week();
+}
+
+if ($result == 1) {
+    \core\notification::success(get_string('bookingconfirmsuccess', 'local_booking', $strdata));
+} elseif ($result == 0) {
+    \core\notification::ERROR(get_string('bookingconfirmunable', 'local_booking'));
+} elseif ($result == -1) {
+    \core\notification::success(get_string('nobookingtoconfirm', 'local_booking'));
+}
 
 if ($result) {
     // redirect