+++ db/messages.php
@@ -79,6 +79,14 @@ $messageproviders = array(
         ),
     ),
 
+    // Notify student of being reinstated after suspension
+    'reinstatement_instructor_notification' => array(
+        'capability' => 'local/booking:instructornotification',
+        'defaults' => array(
+            'airnotifier' => MESSAGE_PERMITTED + MESSAGE_DEFAULT_LOGGEDIN + MESSAGE_DEFAULT_LOGGEDOFF,
+        ),
+    ),
+
     // Notify student of being inactive after posting wait period had passed
     'inactive_warning' => array(
         'capability' => 'local/booking:studentnotification',